+++ NetCoreWebApiPoC.Application.Tests/NetCoreWebApiPoC.Application.Tests.csproj
